4.05 - fixed delete item api, added error message for concurrent deletes, styled mobile menu

+++ backend/src/main/java/com/example/backend/service/ItemServiceImpl.java
@@ -80,6 +80,9 @@ public class ItemServiceImpl implements ItemService{
     
 
     public void deleteItem(Long id) {
+        if(!itemRepository.existsById(id)){
+            throw new ItemNotFoundException(id);
+        }
         itemRepository.deleteById(id);
     }
